requirements.txt
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-decouple>=3.8
requests>=2.31.0
loguru>=0.7.0
